requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.32.0
redis==5.0.8
pyjwt[crypto]==2.9.0
sqlalchemy==2.0.35
alembic==1.13.3
bcrypt==4.1.3
passlib[bcrypt]>=1.7.4
python-multipart==0.0.12
pydantic-settings==2.5.2
psycopg2-binary==2.9.9